hermite: Split cubic functions into their own submodule
